desktop_app/requirements.txt
PyQt6>=6.10.0
PyQt6-Charts>=6.6.0
pymongo==4.6.0
cassandra-driver==3.28.0
neo4j==5.14.0
redis==5.0.1
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
python-jose[cryptography]==3.3.0
email-validator==2.1.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pyspark==3.5.0
APScheduler==3.10.4
faker==20.1.0
pyinstaller==6.3.0
